+++ dpnp/special/__init__.py
@@ -41,8 +41,10 @@ alternatively, is implemented through a subset of python calls.
 
 from ._erf import (
     erf,
+    erfc,
 )
 
 __all__ = [
     "erf",
+    "erfc",
 ]